Give gold_ring a CHA +1 haggle bonus, tier-3 chest slot, and its own asset
gold_ring becomes a NetHack-style ring of adornment: a new chaBonus
item field is summed from equipped gear via a generalized
equipped_bonus(inv, stat) helper (was guard-only) into an effective
CHA that widens NPC haggle price bands (deal_half_band_pct etc. now
take i32). gold_ring re-enters dungeon chest loot at chestTier 3 with
a 10% roll, debuting with the future third dungeon.

Replaces the sword.png placeholder icon with its own Meshy.ai model
(HF-hosted via assets.lock), a Cycles-rendered 128x128 icon, and
concept art in doc/images, with provenance recorded in
doc/assets/items.md and doc updates (ITEM_TIERS.md, TODO.md). Adds a
regression test for the widened band with the ring equipped.

+++ server/src/game_state/trading.rs
@@ -61,7 +61,7 @@ impl TraderDef {
         &self,
         kind: DealKind,
         item_def_id: &str,
-        cha: u8,
+        cha: i32,
     ) -> Result<(u32, i32), &'static str> {
         match self {
             TraderDef::Merchant(m) => {
